% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BestStratification.R
\name{BestStratification}
\alias{BestStratification}
\title{Best stratification fitness evaluation function}
\usage{
BestStratification(solution, eval_func_inputs)
}
\arguments{
\item{solution}{the solution to be evaluated}

\item{eval_func_inputs}{specific inputs for best stratification 
(list with sampling frame and precision constraints)}
}
\description{
Fitness evaluation to be used for the optimization of a sampling frame stratification
}
\details{
This function is the one that performs the evaluation of the fitness in the case of 
the optimization of a sampling frame stratification.
The function takes one of the solutions considered at the k-th iteration of the Quantum
Genetic Algorithm, and determines the minimum sample size required to
be compliant with precision constraints on the target variables.

Requires package SamplingStrata to execute functions aggrStrata and Bethel
}
