% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionMinimum.R
\name{FunctionMinimum}
\alias{FunctionMinimum}
\title{Function Minimum fitness evaluation function}
\usage{
FunctionMinimum(solution, y)
}
\arguments{
\item{solution}{the solution to be evaluated}

\item{eval_func_inputs}{specific input for the evaluation function
(y: the vector of values of the y variable)}
}
\value{
the value of the y variable
}
\description{
Fitness evaluation to be used for finding the minimum of a function
}
\details{
This function is the one that performs the evaluation of the fitness in the case of 
searching the minimum of a function.
The function is trivial: takes one of the solutions considered at the k-th iteration 
(the value of the X), and reports the value of the y.
}
