% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{generate_pop}
\alias{generate_pop}
\title{generate_pop}
\usage{
generate_pop(popsize, genomeLength, q_alphabeta, rot, theta, h, qubit_0)
}
\arguments{
\item{popsize}{the number of generated solutions (population) to be evaluated at each iteration
(default is 20)}

\item{genomeLength}{the length of the genome (or chromosome), representing a possible solution}

\item{q_alphabeta}{the array containing the values of the amplitudes of the qubits}

\item{rot}{the rotation array}

\item{theta}{the rotation angle}

\item{h}{the hadamard gate 
to each individual in the population (default is 1/(popsize+1))}

\item{qubit_0}{the array of qubit 0}
}
\value{
A numeric vector giving the best solution obtained by the QGA
}
\description{
Function to initialize first population
}
