% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KnapsackProblem.R
\name{KnapsackProblem}
\alias{KnapsackProblem}
\title{Knapsack Problem fitness evaluation function}
\usage{
KnapsackProblem(solution, eval_func_inputs)
}
\arguments{
\item{solution}{the solution to be evaluated}

\item{eval_func_inputs}{specific inputs for knapsack problem
(list with list of items with corresponding weights and allowed total weight)}
}
\value{
the total number of items the knapsack
}
\description{
Fitness evaluation to be used for the optimization of the knapsack problem
}
\details{
This function is the one that performs the evaluation of the fitness in the case of 
the optimization of a Knapsack Problem.
The function takes one of the solutions considered at the k-th iteration of the Quantum
Genetic Algorithm, and 

1. counts the number of items in knapsack;
2. verifies if the total weights of the items in the knapsack is below the maximum allowed: 
if not, applies a penalty.
}
