% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QGA.R
\name{QGA}
\alias{QGA}
\title{Quantum Genetic Algorithm}
\usage{
QGA(
  popsize = 20,
  generation_max = 200,
  nbitsol,
  Genome,
  Genome_el,
  thetamax = 3.1415926535 * 0.05,
  thetamin = 3.1415926535 * 0.025,
  pop_mutation_rate_max = 1/(popsize + 1),
  pop_mutation_rate_min = 1/(popsize + 1),
  mutation_rate_max = 1/(Genome + 1),
  mutation_rate_min = 1/(Genome + 1),
  mutation_flag = TRUE,
  eval_function,
  eval_func_inputs
)
}
\arguments{
\item{popsize}{the number of generated solutions (population) to be evaluated at each iteration
(default is 20)}

\item{generation_max}{the number of iterations to be performed
(default is 200)}

\item{nbitsol}{the number of possible values contained in each element of the solution 
(it must be nbitsol <= 2^Genome_el)}

\item{Genome}{the length of the genome, i.e. the elements of the chromosome representing an 
individual in the population}

\item{Genome_el}{the length of the single elements in the elements, where each position 
represents a qubit}

\item{thetamax}{the maximum angle (expressed in radiants) to be used when applyin the rotation gate 
(default is pi * 0.05, where pi = 3.1415926535)}

\item{thetamin}{the minimum angle (expressed in radiants) to be used when applyin the rotation gate 
(default is pi * 0.025, where pi = 3.1415926535)}

\item{pop_mutation_rate_max}{maximum mutation rate to be used when applying the X-Pauli gate, applied 
to each individual in the population (default is 1/(popsize+1))}

\item{pop_mutation_rate_min}{minimum mutation rate to be used when applying the X-Pauli gate, applied 
to each individual in the population (default is 1/(popsize+1))}

\item{mutation_rate_max}{maximum mutation rate to be used when applying the X-Pauli gate, applied 
to each element of the chromosome  (default is 1/(Genome+1)))}

\item{mutation_rate_min}{minimum mutation rate to be used when applying the X-Pauli gate, applied 
to each element of the chromosome (default is 1/(Genome+1))}

\item{mutation_flag}{flag indicating if the mutation gate is to be applied or not (default is TRUE)}

\item{eval_func_inputs}{specific inputs required by the eval_fitness function}

\item{eval_fitness}{name of the function that will be used to evaluate the fitness of each solution}
}
\value{
A numeric vector giving the best solution obtained by the QGA
}
\description{
Main function to execute a Quantum Genetic Algorithm
}
\details{
This function is the 'engine', which performs the quantum genetic algorithm calling
two more functions that need to be developed or personalized:
1. eval_fitness
2. repair_function
}
