% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TravellerSalesman.R
\name{TravellerSalesman}
\alias{TravellerSalesman}
\title{Traveller Salesman Problem fitness evaluation function}
\usage{
TravellerSalesman(solution, distance)
}
\arguments{
\item{solution}{the solution to be evaluated}

\item{eval_func_inputs}{specific input for the Traveler Salesman Problem
(matrix of the distances between the cities)}
}
\value{
the sum of distances
}
\description{
Fitness evaluation to be used for the optimization of a Traveller Salesman Problem
}
\details{
This function is the one that performs the evaluation of the fitness in the case of 
the optimization of a Traveler Salesman Problem.
The function takes one of the solutions considered at the k-th iteration of the Quantum
Genetic Algorithm, and determines the minimum sample size required to
be compliant with precision constraints on the target variables.

Input required is the dataframe containing, for each city, their geographical coordinates.
}
