% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_stratification.R
\name{best_stratification}
\alias{best_stratification}
\title{Best stratification fitness evaluation function}
\usage{
best_stratification(chromosome, generation, eval_func_inputs)
}
\arguments{
\item{chromosome}{the current set of the solutions to be evaluated}

\item{generation}{index of the current generation}

\item{eval_func_inputs}{specific inputs for best stratification 
(list with sampling frame and precision constraints)}
}
\description{
Fitness evaluation to be used for the optimization of a sampling frame stratification
}
\details{
This function is the one that perform the evaluation of the fitness in the case of 
the optimization of a sampling frame stratification.
The function takes the set of solution considered at the k-th iteration of the Quantum
Genetic Algorithm, and for each of them determines the minimum sample size required to
be compliant with precision constraints on the target variables.

Requires package SamplingStrata to execute functions aggrStrata and Bethel
}
