% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionMinimum.R
\name{FunctionMinimum}
\alias{FunctionMinimum}
\title{Function Minimum fitness evaluation function}
\usage{
FunctionMinimum(solution, inputs)
}
\arguments{
\item{solution}{the solution to be evaluated}

\item{eval_func_inputs}{specific inputs for the evaluation function
(n: the number of digits with required by the integer numbers of the x variable,
 y: the vector of values of the y variable)}
}
\description{
Fitness evaluation to be used for finding the minimum of a function
}
\details{
This function is the one that performs the evaluation of the fitness in the case of 
searching the minimum of a function.
The function takes one of the solutions considered at the k-th iteration of the Quantum
Genetic Algorithm, and 

1. convert the vector of binary values representing the current solution (value of the x);
2. calculates
}
